make git save commit messages more informative
Instead of just having "Created by subsurface <version>", put the number
of dives and the location of the last dive in the message.  That makes
things like "gitk" show a much more useful view of what actually got
saved.

We still save the subsurface version in the body of the message, because
that is interesting and relevant information.  It's just not the
*primary* relevant information.

Anyway, with this, a git commit message might looke something like

    dive 474: North West Point (Christmas Island)

    Created by subsurface 4.0.96-17-g649e9ed89d9d

which is much more relevant for the common case of adding new dives at
the end.

Of course, if the reason for the save is that you edited old dives, the
relevance of the commit message telling you the number of dives you have
in the log and the dive number is questionable.  But then you have to
look at the actual diff to see what's going on.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>