git load: initialize dive computer timestamp and duration from dive
We don't save the dive computer timestamp and duration if they match the
dive timestamp and duration.  But that means that on loading, we need to
default the dive computer time/duration to the dive one.  If they
differ, the loading of the divecomputer file will then override the
default timestamp/duration.

This mainly matters if a later dive merge then changes the timestamp of
the dive: the dive computer timestamp needs to have been set correctly
and not change.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>